Create Create An easy way to reduce the length of a link
This Python script uses the pyshorteners library to shorten URLs using the TinyURL service. It prompts the user to input a long URL, then uses pyshorteners.Shortener() to create a shortener object. The short() method of the TinyURL service is called on the object to generate a shortened URL, which is then printed to the console.
Notes:
Make sure you have the pyshorteners library installed (pip install pyshorteners) before running the script.
This script demonstrates a simple yet practical use of URL shortening services, which can be useful in various applications, including sharing links via social media or reducing URL length in communications.